[QI2-865] remove old circuit.py and leave v3 version, renamed
